// lib/components/TextImage/TextImage.Styles.d.ts
export declare const styles: {
    base: string;
    baseContent: string;
    border: string;
    headerOne: string;
    headerTwo: string;
    underline: string;
    contentWrapper: string;
    pronoun: string;
    listLink: string;
};
export declare const mediaStyles: {
    mediaWrapper: string;
    mediaBgImage: string;
    svgLeft: string;
    svgRight: string;
    svgFlipLeft: string;
    svgFlipRight: string;
};
export declare const socialStyles: {
    flex: string;
    link: string;
    icon: string;
};